requirements.txt
APScheduler==3.10.4
fastapi==0.111.0
fastapi-mail==1.4.1
pyhumps==3.8.0
pydantic==2.7.3
pydantic_settings==2.2.1
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
SQLAlchemy==2.0.30
starlette==0.37.2
psycopg2-binary==2.9.9